Move components from claim-conditions.tsx to separate files
